requirements.txt
acme==0.31.0
asgiref==3.2.3
asn1crypto==0.24.0
beautifulsoup4==4.8.2
blinker==1.4
certbot==0.31.0
certbot-nginx==0.31.0
certifi==2020.6.20
chardet==3.0.4
cloud-init==19.2
command-not-found==0.3
ConfigArgParse==0.11.0
configobj==5.0.6
cryptography==2.1.4
distro-info===0.18ubuntu0.18.04.1
Django==3.0.3
django-cors-headers==3.2.1
django-debug-panel==0.8.3
django-debug-toolbar==2.2
djangorestframework==3.11.0
dnspython==1.16.0
future==0.15.2
idna==2.10
Jinja2==2.10
josepy==1.1.0
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
MarkupSafe==1.0
mock==2.0.0
mysql-connector-python==8.0.19
ndg-httpsclient==0.4.4
netifaces==0.10.4
oauthlib==2.0.6
parsedatetime==2.4
pbr==3.1.1
protobuf==3.6.1
psycopg2==2.8.4
pyasn1==0.4.2
pycrypto==2.6.1
pydub==0.23.1
pygobject==3.26.1
PyICU==1.9.8
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyRFC3339==1.0
pyserial==3.4
pyTelegramBotAPI==3.7.2
python-apt==1.6.4
python-slugify==4.0.0
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
requests==2.24.0
requests-toolbelt==0.8.0
SecretStorage==2.3.1
six==1.15.0
soupsieve==1.9.5
sqlparse==0.3.0
ssh-import-id==5.7
text-unidecode==1.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.10
uWSGI==2.0.18
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.3.2
